DATA605/Spring2025/projects/TutorTask56_Spring2025_Real-time_Bitcoin_Data_Analysis_with_Snowflake/requirements.txt
pandas
requests
python-dotenv
snowflake-connector-python
jupyterlab
plotly
boto3
dotenv
cryptography
Prophet
streamlit